Feature(Modularized Edit and Graph Views)
